+++ .github/workflows/checks.yml
@@ -66,7 +66,13 @@ jobs:
         uses: actions-rs/cargo@v1.0.3
         with:
           command: rustdoc
-          args: -p taskchampion-sync-server --all-features -- -Z unstable-options  --check -Dwarnings
+          args: -p taskchampion-sync-server --bin taskchampion-sync-server --all-features -- -Z unstable-options  --check -Dwarnings
+
+      - name: taskchampion-sync-server-postgres
+        uses: actions-rs/cargo@v1.0.3
+        with:
+          command: rustdoc
+          args: -p taskchampion-sync-server --bin taskchampion-sync-server-postgres --all-features -- -Z unstable-options  --check -Dwarnings
 
       - name: taskchampion-sync-server-core
         uses: actions-rs/cargo@v1.0.3
@@ -79,6 +85,12 @@ jobs:
         with:
           command: rustdoc
           args: -p taskchampion-sync-server-storage-sqlite --all-features -- -Z unstable-options  --check -Dwarnings
+
+      - name: taskchampion-sync-server-storage-postgres
+        uses: actions-rs/cargo@v1.0.3
+        with:
+          command: rustdoc
+          args: -p taskchampion-sync-server-storage-postgres --all-features -- -Z unstable-options  --check -Dwarnings
 
   fmt:
     runs-on: ubuntu-latest
