+++ server/src/api/get_snapshot.rs
@@ -34,8 +34,10 @@ pub(crate) async fn service(
 
 #[cfg(test)]
 mod test {
-    use crate::WebServer;
-    use crate::{api::CLIENT_ID_HEADER, WebConfig};
+    use crate::{
+        api::CLIENT_ID_HEADER,
+        web::{WebConfig, WebServer},
+    };
     use actix_web::{http::StatusCode, test, App};
     use chrono::{TimeZone, Utc};
     use pretty_assertions::assert_eq;
